Update README.md
Nowadays, young people are struggling to know what exactly they want to do after highschool. Because of that, most youth seek the internet to help them discover what career they would be best fitted in. However, due to the vast amount of information, it can be overwhelming and confusing to really comprehend this information. To solve this, we will design an gamiefied career guidance app that, based on selection, assists users to understand what career would best suit their personality types.  This is an easier way for young students to  get the information they need in a fun interactive gamified environment.